Moved the assertCanAppend method to the shoppingCart test setup to be able to reuse it
