+++ src/packages/emmett-mongodb/src/eventStore/projections/mongoDBEventStore.projections.e2e.spec.ts
@@ -26,7 +26,7 @@ import {
   type PricedProductItem,
   type ProductItemAdded,
   type ShoppingCartEvent,
-} from '../../testing/shoppingCart.domain';
+} from '../../testing';
 
 void describe('MongoDBEventStore', () => {
   let mongodb: StartedMongoDBContainer;
