+++ src/packages/emmett-mongodb/src/testing/index.ts
@@ -1,2 +1,2 @@
-export * from './mongoMocks';
+export * from './mongoDummies';
 export * from './shoppingCart.domain';
