+++ src/packages/emmett-mongodb/src/testing/shoppingCart.domain.ts
@@ -1,4 +1,16 @@
-import { type Event, type ReadEvent } from '@event-driven-io/emmett';
+import {
+  assertNotEqual,
+  STREAM_DOES_NOT_EXIST,
+  type Event,
+  type ReadEvent,
+} from '@event-driven-io/emmett';
+import { v7 as uuid } from 'uuid';
+import {
+  MongoDBEventStoreDefaultStreamVersion,
+  toStreamName,
+  type MongoDBEventStore,
+  type StreamType,
+} from '../eventStore';
 
 export type PricedProductItem = {
   productId: string;
@@ -66,4 +78,27 @@ export const evolveWithMetadata = (
 
 export const initialState = (): ShoppingCart => {
   return { productItems: [], totalAmount: 0 };
+};
+
+export const ShoppingCartStreamType: StreamType = 'shopping_cart';
+
+export const assertCanAppend = async (eventStore: MongoDBEventStore) => {
+  const productItem: PricedProductItem = {
+    productId: '123',
+    quantity: 10,
+    price: 3,
+  };
+  const shoppingCartId = uuid();
+  const streamName = toStreamName(ShoppingCartStreamType, shoppingCartId);
+
+  const result = await eventStore.appendToStream<ShoppingCartEvent>(
+    streamName,
+    [{ type: 'ProductItemAdded', data: { productItem } }],
+    { expectedStreamVersion: STREAM_DOES_NOT_EXIST },
+  );
+
+  assertNotEqual(
+    result.nextExpectedStreamVersion,
+    MongoDBEventStoreDefaultStreamVersion,
+  );
 };
